refactor: move encrytion verifiation process into utils file
